+++ blazeweb/__init__.py
@@ -1,1 +1,4 @@
-VERSION = '0.4.9'
+from os import path as osp
+
+cdir = osp.abspath(osp.dirname(__file__))
+VERSION = open(osp.join(cdir, 'version.txt')).read().strip()